+++ R/axistetra.R
@@ -10,18 +10,24 @@
 #' (applied for x, y and z) or a vector of 3 separate values for each axis.
 #' @param arrowhead size of the arrowhead.
 #' @param col,lty,lwd graphical parameters for the arrows.
-#' @param label logical, include x, y and z labels (defaults to TRUE).
+#' @param label logical, include x, y and z labels (defaults to `TRUE`).
 #' @param adj.label position adjustment for the labels. a list of 3 named objects
 #' for x, y and z arrows, each with 2 values for x and y adjustment.
 #' @param label.cex,label.col graphical parameters for the labels.
 #'
-#' @return \code{axistetra} adds reference arrows showing the direction of the
+#' @return `axistetra` adds reference arrows showing the direction of the
 #' 3-dimensional axes in a static tetrahedral colourspace plot.
 #'
 #' @author Rafael Maia \email{rm72@@zips.uakron.edu}
 #'
 #' @export
-
+#'
+#' @examples
+#' data(sicalis)
+#' vis.sicalis <- vismodel(sicalis, visual = "avg.uv")
+#' tcs.sicalis <- colspace(vis.sicalis, space = "tcs")
+#' plot(tcs.sicalis)
+#' axistetra()
 axistetra <- function(x = 0, y = 1.3, size = 0.1,
                       arrowhead = 0.05, col = par("fg"),
                       lty = par("lty"), lwd = par("lwd"),