+++ R/coldist.R
@@ -1,95 +1,70 @@
 #' Colour distances
 #'
-#' Calculates colour distances. When data are the result of \code{\link{vismodel}},
+#' Calculates colour distances. When data are the result of [vismodel()],
 #' it applies the receptor-noise model of Vorobyev et al. (1998) to calculate colour distances
-#' with noise based on relative photoreceptor densities. It also accepts \code{\link{colspace}} data
-#' from the hexagon, colour-opponent-coding, categorical, segment, and cielab models, in which case euclidean
-#' distances (hexagon, cielab, categorical, segment) or manhattan distances (coc) are returned.
+#' with noise based on relative photoreceptor densities. It also accepts [colspace()] data in which case
+#' unweighted Euclidean distances or Manhattan distances (coc model only) are returned.
 #'
 #' @param modeldata (required) quantum catch colour data. Can be the result
-#'  from \code{\link{vismodel}}, or \code{\link{colspace}}. Data may also be independently calculated quantum catches,
-#'  in the form of a data frame with columns representing photoreceptors.
-#' @param qcatch if the object is of class \code{vismodel} or \code{colspace},
+#'  from [vismodel()] for noise-weighted Euclidean distances, or [colspace()] for
+#'  unweighted (typically) Euclidean distances. Data may also be independently calculated
+#'  quantum catches, in the form of a data frame with columns representing photoreceptors.
+#' @param qcatch if the object is of class [`vismodel`] or [`colspace`],
 #'  this argument is ignored. If the object is a data frame of quantal catches
 #'  from another source, this argument is used to specify what type of quantum catch is being
 #'  used, so that the noise can be calculated accordingly:
-#' \itemize{
-#'  \item \code{Qi}: Quantum catch for each photoreceptor
-#'  \item \code{fi}: Quantum catch according to Fechner's law (the signal of the receptor
-#'    channel is proportional to the logarithm of the quantum catch)
-#'  }
+#'  * `Qi`: Quantum catch for each photoreceptor
+#'  * `fi`: Quantum catch according to Fechner's law (the signal of the receptor
+#'  channel is proportional to the logarithm of the quantum catch)
 #' @param subset If only some of the comparisons should be returned, a character vector of
 #'  length 1 or 2 can be provided, indicating which samples are desired. The subset vector
 #'  must match the labels of the input samples, but partial matching (and regular expressions)
 #'  are supported.
-#' @param achromatic Logical. If \code{TRUE}, last column of the data frame is used to calculate
-#'  the achromatic contrast, with noise based on the Weber fraction given by the argument
-#'  \code{weber.achro}.
-#'  If the data are from the hexagon model (i.e. \code{colspace(space = 'hexagon')}), it
-#'  instead returns simple long (or 'green') receptor contrast.
+#' @param achromatic Logical. If `TRUE`, last column of the data frame is used to calculate
+#'  the achromatic contrast, the form of which will depend on the input data and will be
+#'  indicated by a message during execution. For noise-weighted distances, noise is based on
+#'  the Weber fraction given by the argument `weber.achro`.
 #' @param n photoreceptor densities for the cones used in visual modeling.
 #'  must have same length as number of columns (excluding achromatic receptor if used;
-#'  defaults to
-#'  the Pekin robin \emph{Leiothrix lutea} densities: \code{c(1,2,2,4)}).
-#'  Ignored for \code{colspace} objects
-#'  if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding,
-#'  categorical, and cie models).
+#'  defaults to the Pekin robin *Leiothrix lutea* densities: `c(1,2,2,4)`).
+#'  Ignored for [`colspace`] objects.
 #' @param weber The Weber fraction to be used (often also referred to as receptor noise,
-#'  or \emph{e}). The noise-to-signal ratio \code{v} is unknown,
+#'  or *e*). The noise-to-signal ratio `v` is unknown,
 #'  and therefore must be calculated based on the empirically estimated Weber
-#'  fraction of one of the cone classes. \code{v} is then applied to estimate the
+#'  fraction of one of the cone classes. `v` is then applied to estimate the
 #'  Weber fraction of the other cones. by default, the value of 0.1 is used
 #'  (the empirically estimated value for the
-#'  LWS cone from \emph{Leiothrix lutea}). See Olsson et al. 2017 for a review of
-#'  published values in the literature. Ignored for \code{colspace} objects
-#'  if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding,
-#'  categorical, segment, and cie models).
+#'  LWS cone from *Leiothrix lutea*). See Olsson et al. 2017 for a review of
+#'  published values in the literature. Ignored for `colspace` objects.
 #' @param weber.ref the cone class used to obtain the empirical estimate of the
-#'  Weber fraction used for the \code{weber} argument. By default, \code{n4} is used,
-#'  representing the LWS cone for \emph{Leiothrix lutea}. Ignored for \code{colspace} objects
-#'  if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding,
-#' categorical, segment, and cie models).
+#'  Weber fraction used for the `weber` argument. By default, `n4` is used,
+#'  representing the LWS cone for *Leiothrix lutea*. Ignored for `colspace` objects.
 #' @param weber.achro the Weber fraction to be used to calculate achromatic contrast, when
-#'  \code{achromatic = TRUE}. Defaults to 0.1. Ignored for \code{colspace} objects
-#'  if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding,
-#' categorical, segment, and cie models).
-#' @param noise how the noise will be calculated. (Ignored for \code{colspace} objects
-#'  if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding,
-#' categorical, segment, and cie models)):
-#' \itemize{
-#' 	\item \code{neural} (default): noise is proportional to the Weber fraction and
+#'  `achromatic = TRUE`. Defaults to 0.1. Ignored for `colspace` objects.
+#' @param noise how the noise will be calculated (ignored for `colspace` objects):
+#' * `neural` (default): noise is proportional to the Weber fraction and
 #'  is independent of the intensity of the signal received (i.e. assumes bright conditions).
-#' 	\item \code{quantum}: noise is the sum of the neural noise and receptor noise,
+#' * `quantum`: noise is the sum of the neural noise and receptor noise,
 #'  and is thus proportional to the Weber fraction and inversely proportional
 #'  to the intensity of the signal received (the quantum catches).
-#'  Note that the \code{quantum} option will only work with
-#' 	objects of class \code{vismodel}.
-#' }
+#'  Note that the `quantum` option will only work with
+#' objects of class `vismodel`.
 #'
 #' @return A data frame containing up to 4 columns.
-#' The first two (\code{patch1, patch2}) refer
-#' to the two colors being contrasted; \code{dS} is the chromatic contrast (delta S)
-#' and \code{dL} is the achromatic contrast (delta L). Units are JND's in the receptor-noise
-#' model, euclidean distances in the categorical and segment space, manhattan distances in the
-#' color-opponent-coding space, green-receptor contrast in the hexagon, and lightness (L)
-#' contrast in the cielab model.
+#' The first two (`patch1, patch2`) refer
+#' to the two colors being contrasted; `dS` is the chromatic contrast (delta S)
+#' and `dL` is the achromatic contrast (delta L). Units of `dS` JND's in the receptor-noise
+#' model, unweighted Euclidean distances in colorspace models, and Manhattan distances in the
+#' color-opponent-coding space. Units of `dL` vary, and are either simple contrast, Weber contrast,
+#' or Michelson contrast, as indicated by the output message.
 #'
-#' @section Note on previous versions:
-#' previous versions of \code{coldist} calculated receptor noise using the arguments
-#' \code{v} for the individual cone noise-to-signal ratio and \code{n1,n2,n3,n4} for
-#' the relative cone densities. These arguments have been replaced by \code{weber} and
-#' \code{n}, which takes a vector of relative cone densities. \code{weber.ref} allows
-#' the user to specify which receptor to use as the reference to obtain the
-#' desired Weber fraction, and \code{coldist} calculates internally the value of \code{v}
-#' to be used when calculating the Weber fraction for the remaining cones.
-#'
-#' This allows a more explicit choice of Weber fraction, without the need to find the
-#' right value of \code{v} to use in order to obtain the desired signal-to-noise ratio. Furthermore,
-#' by allowing \code{n} to be entered as a vector, \code{coldist} can now handle visual
-#' systems with more than four photoreceptors.
-#'
-#' In addition, the achromatic noise is calculated based on the \code{weber.achro}
-#' argument directly, and not based on \code{v} and \code{n4} as before.
+#' @section Note on previous versions: Generic di- tri- and tetra-chromatic `colspace()` objects were previously
+#' passed through the receptor-noise limited model to return noise-weighted Euclidean distances. This 
+#' behaviour has been amended, and generic spaces now return unweighted Euclidean distances. Equivalent 
+#' results to the former behaviour can be attained by sending the results of `vismodel()` directly to `coldist()`
+#' , as previously, which also offers greater flexibility and reliability. Thus `coldist()` now returns
+#'  unweighted Euclidean distances for `colspace()` objects (with the exception of Manhattan distances for 
+#'  the coc space), and noise-weighted Euclidean distances for `vismodel()` objects.   
 #'
 #' @export
 #'
@@ -99,36 +74,31 @@
 #' data(flowers)
 #' vis.flowers <- vismodel(flowers, visual = "canis", relative = FALSE)
 #' didist.flowers <- coldist(vis.flowers, n = c(1, 2))
-#' 
+#'
 #' # Trichromat
 #' vis.flowers <- vismodel(flowers, visual = "apis", relative = FALSE)
 #' tridist.flowers <- coldist(vis.flowers, n = c(1, 2, 1))
-#' 
-#' # Trichromat, color-hexagon model (euclidean distances)
+#'
+#' # Trichromat, colour-hexagon model (euclidean distances)
 #' vis.flowers <- vismodel(flowers,
 #'   visual = "apis", qcatch = "Ei",
 #'   relative = FALSE, vonkries = TRUE, achro = "l", bkg = "green"
 #' )
 #' hex.flowers <- colspace(vis.flowers, space = "hexagon")
 #' hexdist.flowers <- coldist(hex.flowers)
-#' 
-#' # Trichromat, color-opponent-coding model (manhattan distances)
+#'
+#' # Trichromat, colour-opponent-coding model (manhattan distances)
 #' vis.flowers <- vismodel(flowers, visual = "apis", qcatch = "Ei", relative = FALSE, vonkries = TRUE)
 #' coc.flowers <- colspace(vis.flowers, space = "coc")
 #' hexdist.flowers <- coldist(coc.flowers)
-#' 
+#'
 #' # Tetrachromat
 #' data(sicalis)
 #' vis.sicalis <- vismodel(sicalis, visual = "avg.uv", relative = FALSE)
 #' tetradist.sicalis.n <- coldist(vis.sicalis)
 #' 
-#' # This will also work, but give you several warnings you shouldn't ignore!!
-#' col.sicalis <- colspace(vis.sicalis)
-#' tetradist.sicalis.n <- coldist(col.sicalis)
-#' 
-#' tetradist.sicalis.q <- coldist(vis.sicalis, noise = "quantum")
 #' }
-#' 
+#'
 #' @author Rafael Maia \email{rm72@@zips.uakron.edu}
 #'
 #' @references Vorobyev, M., Osorio, D., Bennett, A., Marshall, N., & Cuthill, I.
@@ -169,7 +139,7 @@ coldist <- function(modeldata,
 
   noise <- match.arg(noise)
 
-  usereceptornoisemodel <- !isTRUE(attr(modeldata, "clrsp") %in% c("hexagon", "categorical", "CIELAB", "CIELch", "segment", "coc"))
+  usereceptornoisemodel <- !isTRUE(any(class(modeldata) %in% "colspace"))
 
   if (noise == "quantum") {
     if (!is.vismodel(modeldata) && !is.colspace(modeldata)) {
@@ -188,21 +158,24 @@ coldist <- function(modeldata,
   # Pre-processing for colspace objects
   if (is.colspace(modeldata) || is.vismodel(modeldata)) {
     qcatch <- attr(modeldata, "qcatch")
-  }
-
-  # Pre-processing for vismodel objects
-  if (is.vismodel(modeldata)) {
-    # Set achromatic=FALSE if visual model has achromatic='none'
-    if (attr(modeldata, "visualsystem.achromatic") == "none" && achromatic) {
-      warning("achromatic=TRUE but visual model was calculated with achromatic=",
-        dQuote("none"), "; achromatic contrast not calculated.",
-        call. = FALSE
-      )
-      achromatic <- FALSE
+    # Pre-processing for vismodel objects
+    if (is.vismodel(modeldata)) {
+      if (qcatch == "Ei") {
+        stop("Receptor-noise model not compatible with hyperbolically transformed quantum catches (Ei)", call. = FALSE)
+      }
     }
-    # initial checks...
-    if (qcatch == "Ei") {
-      stop("Receptor-noise model not compatible with hyperbolically transformed quantum catches (Ei)", call. = FALSE)
+    # Convert lum values to 0 instead of NA, for convenient
+    # processing. Converted back to NA at the end.
+    if (attr(modeldata, "visualsystem.achromatic") == "none" && !(any(c("CIELAB", "CIELCh") %in% attr(modeldata, "clrsp")))
+        || is.null(attr(modeldata, "visualsystem.achromatic"))) {
+      modeldata$lum <- 0
+      if (achromatic) {
+        warning("achromatic = TRUE but visual model was calculated with achromatic = ",
+          dQuote("none"), "; achromatic contrast not calculated.",
+          call. = FALSE
+        )
+      }
+      achromatic <- FALSE
     }
   }
 
@@ -216,7 +189,12 @@ coldist <- function(modeldata,
       )
     }
     else {
-      ncone <- ncol(modeldata)
+      # Don't count all-NA columns when guessing ncone
+      if (any(sapply(modeldata, function(x) all(is.na(x))))) {
+        ncone <- ncol(modeldata) - 1
+      } else {
+        ncone <- ncol(modeldata)
+      }
       warning("number of cones not specified; assumed to be ", ncone,
         call. = FALSE
       )
@@ -230,24 +208,22 @@ coldist <- function(modeldata,
   }
 
   if (usereceptornoisemodel) {
+
     #########################
     # Receptor Noise Models #
     #########################
 
     # should be used when:
-    # - colspace object: is not hexagon, coc, categorical, ciexyz, cielab, cielch
+    # - colspace object: never
     # - vismodel object: always
     # - user input data: always
 
-
-    if (any(c("dispace", "trispace", "tcs") %in% attr(modeldata, "clrsp"))) {
-      dat <- as.matrix(modeldata[, names(modeldata) %in% c("u", "s", "m", "l", "lum")])
-    } else {
-      dat <- as.matrix(modeldata)
-
-      rownames(dat) <- rownames(modeldata)
-      colnames(dat) <- colnames(modeldata)
-    }
+    note_dS <- 'Calculating noise-weighted Euclidean distances'
+    note_dL <- NULL
+    
+    dat <- as.matrix(modeldata)
+    rownames(dat) <- rownames(modeldata)
+    colnames(dat) <- colnames(modeldata)
 
     # Ensure catches are log transformed
     dat <- switch(qcatch,
@@ -325,6 +301,7 @@ coldist <- function(modeldata,
     )
 
     if (achromatic) {
+      note_dL <- ' and noise-weighted luminance contrasts'
       visref <- cbind(visref, lum = log(1e-10))
       visref[grep("jnd2xyzrrf", rownames(visref), invert = TRUE), "lum"] <-
         dat[seq(refsamp), dim(dat)[2]]
@@ -360,27 +337,61 @@ coldist <- function(modeldata,
         warning("achromatic is set to TRUE, but input data has the same number of columns for sensory data as number of cones in the visual system. There is no column in the data that represents an exclusively achromatic channel, last column of the sensory data is being used. Treat achromatic results with caution, and check if this is the desired behavior.", call. = FALSE)
       }
     }
+    message(note_dS, note_dL)
   } else {
     dat <- as.matrix(modeldata[, sapply(modeldata, is.numeric)])
 
-    res[, "dS"] <- switch(attr(modeldata, "clrsp"),
+    # Message about the distances being calculated
+    note_dS <- switch(attr(modeldata, "clrsp"),
+      "dispace" = ,
+      "trispace" = ,
+      "tcs" = ,
       "hexagon" = ,
-      "categorical" = apply(pairsid, 1, function(x) euc(dat[x[1], c("x", "y")], dat[x[2], c("x", "y")])),
+      "categorical" = ,
+      "segment" = "Calculating unweighted Euclidean distances",
       "CIELAB" = ,
-      "CIELch" = apply(pairsid, 1, function(x) euc(dat[x[1], c("L", "a", "b")], dat[x[2], c("L", "a", "b")])),
-      "segment" = apply(pairsid, 1, function(x) euc(dat[x[1], c("MS", "LM")], dat[x[2], c("MS", "LM")])),
-      "coc" = apply(pairsid, 1, function(x) bloc2d(dat[x[1], ], dat[x[2], ]))
+      "CIELCh" = "Calculating CIE2000 distances",
+      "coc" = "Calculating Manhattan distances"
+    )
+    note_dL <- NULL
+
+    res[, "dS"] <- switch(attr(modeldata, "clrsp"),
+      "dispace" = apply(pairsid, 1, function(x) dist(rbind(dat[x[1], "x"], dat[x[2], "x"]))),
+      "tcs" = apply(pairsid, 1, function(x) dist(rbind(dat[x[1], c("x", "y", "z")], dat[x[2], c("x", "y", "z")]))),
+      "trispace" = ,
+      "hexagon" = ,
+      "categorical" = apply(pairsid, 1, function(x) dist(rbind(dat[x[1], c("x", "y")], dat[x[2], c("x", "y")]))),
+      "segment" = apply(pairsid, 1, function(x) dist(rbind(dat[x[1], c("MS", "LM")], dat[x[2], c("MS", "LM")]))),
+      "CIELAB" = ,
+      "CIELCh" = apply(pairsid, 1, function(x) dist(rbind(dat[x[1], c("L", "a", "b")], dat[x[2], c("L", "a", "b")]))),
+      "coc" = apply(pairsid, 1, function(x) dist(rbind(dat[x[1], c("x", "y")], dat[x[2], c("x", "y")]), method = "manhattan"))
     )
     if (achromatic) {
-      res[, "dL"] <- switch(attr(modeldata, "clrsp"),
-        "hexagon" = apply(pairsid, 1, function(x) achrohex(dat[x[1], ], dat[x[2], ])),
-        "categorical" = NA,
+      note_dL <- switch(attr(modeldata, "clrsp"),
+        "dispace" = ,
+        "trispace" = ,
+        "tcs" = " and Weber luminance contrast",
+        "hexagon" = " and simple luminance contrast",
+        "segment" = " and Michelson luminance contrast",
         "CIELAB" = ,
-        "CIELch" = apply(pairsid, 1, function(x) euc(dat[x[1], "L"], dat[x[2], "L"])),
-        "segment" = apply(pairsid, 1, function(x) euc(dat[x[1], "B"], dat[x[2], "B"])),
+        "CIELCh" = " and Weber luminance contrast",
+        "categorical" = ,
+        "coc" = " and no luminance contrast"
+      )
+
+      res[, "dL"] <- switch(attr(modeldata, "clrsp"),
+        "dispace" = ,
+        "trispace" = ,
+        "tcs" = apply(pairsid, 1, function(x) lumcont(dat[x[1], "lum"], dat[x[2], "lum"], type = "weber")),
+        "categorical" = NA,
+        "hexagon" = apply(pairsid, 1, function(x) lumcont(dat[x[1], "l"], dat[x[2], "l"], type = "simple")),
+        "CIELAB" = ,
+        "CIELCh" = apply(pairsid, 1, function(x) lumcont(dat[x[1], "L"], dat[x[2], "L"], type = "weber")),
+        "segment" = apply(pairsid, 1, function(x) lumcont(dat[x[1], "B"], dat[x[2], "B"], type = "michelson")),
         "coc" = NA
       )
     }
+    message(note_dS, note_dL)
   }
 
   # Subsetting samples
@@ -413,6 +424,12 @@ coldist <- function(modeldata,
 
   if (exists("resref", inherits = FALSE)) {
     attr(res, "resref") <- resref
+  }
+
+  # Set achro contrasts to NA if no lum values supplied
+  if (attr(modeldata, "visualsystem.achromatic") == "none" || is.null(attr(modeldata, "visualsystem.achromatic"))
+  || !(achromatic)) {
+    res$dL <- NA
   }
 
   attr(res, "ncone") <- ncone
@@ -522,20 +539,16 @@ ttdistcalcachro <- function(f1, f2, qn1 = NULL, qn2 = NULL, weber.achro) {
 # START OTHER DISTANCES #
 #########################
 
+# Luminance contrast
+lumcont <- function(coord1, coord2, type = c("simple", "weber", "michelson")) {
+  contrast <- match.arg(type)
 
-# Euclidean distance
-euc <- function(coord1, coord2) {
-  sqrt(sum((coord1 - coord2)^2))
-}
-
-# Achromatic 'green' receptor contrast in the hexagon
-achrohex <- function(coord1, coord2) {
-  coord1["l"] / coord2["l"]
-}
-
-# Manhattan distance
-bloc2d <- function(coord1, coord2) {
-  abs(coord1["x"] - coord2["x"]) + abs(coord1["y"] - coord2["y"])
+  dLout <- switch(contrast,
+    "simple" = coord1 / coord2,
+    "weber" = (max(c(coord1, coord2)) - min(c(coord1, coord2))) / min(c(coord1, coord2)),
+    "michelson" = (max(c(coord1, coord2)) - min(c(coord1, coord2))) / (max(c(coord1, coord2)) + min(c(coord1, coord2)))
+  )
+  dLout
 }
 
 # CIE2000 colour distance for CIELCh (LOLWAT)