+++ R/irrad2flux.R
@@ -1,13 +1,13 @@
 #' Converts between irradiance and photon (quantum) flux
 #'
 #' Some spectrometers will give illuminant values in units of irradiance
-#' (uWatt * cm^-2), but physiological models require illuminants in units of
-#' photon (quantum) flux (umol * s^-1 * m^-2). The functions \code{irrad2flux}
-#' and \code{flux2irrad} allows for easy conversion of \code{rspec} objects
+#' (\eqn{\mu Watt.cm^{-2}}), but physiological models require illuminants in units of
+#' photon (quantum) flux (\eqn{\mu mol.s^{-1}.m^{-2}}). The functions `irrad2flux`
+#' and `flux2irrad` allows for easy conversion of `rspec` objects
 #' between these units.
 #'
 #' @param rspecdata (required) a rspec object containing illuminant values.
-#' @return a converted \code{rspec} object.
+#' @return a converted `rspec` object.
 #' @export irrad2flux flux2irrad
 #'
 #' @author Rafael Maia \email{rm72@@zips.uakron.edu}
@@ -20,7 +20,7 @@ irrad2flux <- function(rspecdata) {
   nam <- names(rspecdata)
   wl <- rspecdata[, nam == "wl"]
 
-  K <- 0.01 / (6.626 * 2.998 * 6.02308)
+  K <- 0.01 / (6.626068 * 2.99792458 * 6.02214076)
 
   rspecdata[, nam != "wl"] <- rspecdata[, nam != "wl"] * wl * K
 
@@ -38,7 +38,7 @@ flux2irrad <- function(rspecdata) {
   nam <- names(rspecdata)
   wl <- rspecdata[, nam == "wl"]
 
-  K <- 0.01 / (6.626 * 2.998 * 6.02308)
+  K <- 0.01 / (6.626068 * 2.99792458 * 6.02214076)
 
   rspecdata[, nam != "wl"] <- rspecdata[, nam != "wl"] / (wl * K)