+++ tests/testthat/test-general.R
@@ -1,4 +1,3 @@
-library(pavo)
 context("general")
 
 test_that("Class assignment", {
@@ -26,7 +25,9 @@ test_that("Class assignment", {
 })
 
 test_that("sensdata", {
+  library(digest)
   expect_true(all(names(as.data.frame(vissyst)) %in% names(sensdata("all", "all"))))
+  expect_equal(digest::sha1(sensdata(illum = 'all', bkg = 'all', trans = 'all'), digits = 4), "581d632a210f2dc51e19034d543854e5e2a9056e")
 })
 
 test_that("peakshape", {
@@ -36,4 +37,23 @@ test_that("peakshape", {
 
   test <- read.csv("known_output/FWHM_lims.csv")
   expect_equal(peakshape(test, plot = FALSE)[, 4], c(144, 52))
+
+  expect_warning(peakshape(flowers[, -1], plot = FALSE), "No wavelengths")
+
+  expect_equivalent(
+    nrow(peakshape(flowers, grepl("^Hibbertia", colnames(flowers)), plot = FALSE)),
+    6
+  )
+
+  expect_null(peakshape(flowers, select = FALSE))
+
+  expect_equivalent(
+    digest::sha1(peakshape(flowers, absolute.min = TRUE), digits = 5),
+    "5300bb69c576646c251857257b265972c29536e2"
+  )
+
+  expect_warning(
+    peakshape(flowers, lim = c(300, 400), plot = FALSE),
+    "incorporate all minima in spectral curves"
+  )
 })