+++ NEWS.md
@@ -1,3 +1,24 @@
+# pavo 2.2.0
+
+## NEW FEATURES AND SIGNIFICANT CHANGES
+
+* added k-medoids clustering as an option in `classify()`
+* `coldist()` now returns unweighted Euclidean distances for generic di-, tri-, and tetra-chromatic `colspace()` models. Users wishing to estimate noise-weighted distances (i.e. via the receptor-noise limited model) should continue to use `coldist()` directly on `vismodel()` objects, as previously.
+
+## MINOR FEATURES AND BUG FIXES
+
+* added a message to indicate what distance metric is being calculated during calls to `coldist()` 
+* `vismodel()`, `colspace()`, and `coldist()` now always return `lum` and `dL` columns, for consistency
+* `irrad2flux()` and `flux2irrad()` use slightly more precise constant values, which may lead to slightly different results (less than 0.1% difference between this new version and the previous versions)
+* `getimg()` can now import images from URL's
+* the `cores` argument in `getimg()` is deprecated, as image import is vectorised
+* fixed a small bug in `classify()` when using `refID` and `kcols` for multiple images
+* the `resize` argument in `procimg()` now takes a percentage, rather than scaling factor
+* fixed a bug where `colspace(space = segment)` would return both a `B` and (redundant) `lum` column
+* fixed a bug where the rod sensitivity of _Canis familiaris_ was inaccessible through `vismodel()`
+* fixed an issue in `getspec()` where badly encoded characters in some spectral files would cause failure
+* fixed a bug where `coldist()` would attempt to estimate receptor-noise weigted distances rather than cie-distances for `cielch` model results
+
 # pavo 2.1.0
 
 ## NEW FEATURES AND POTENTIALLY BREAKING CHANGES